Fixed stamina rounding error bug

+++ Server/Config/TimeBasedSettings.cs
@@ -3,7 +3,7 @@
     public class TimeBasedSettings
     {
         public int StaminaRegenAmount { get; set; }
-        public int StaminaRegenIntervalHours { get; set; }
+        public double StaminaRegenIntervalHours { get; set; }
         public int HungerDepletionAmount { get; set; }
         public int HungerDepletionIntervalHours { get; set; }
         public double HungerDepletionMinThreshold { get; set; }
